refactor: remove inappropriate scenarios from ScenarioFeature
