+++ src/test/Xbehave.Features.Net40/Xbehave.Features.Net40.csproj
@@ -71,6 +71,7 @@
     <Compile Include="Infrastructure\Pass.cs" />
     <Compile Include="Infrastructure\Result.cs" />
     <Compile Include="ContinueOnFailureFeature.cs" />
+    <Compile Include="DefaultParametersFeature.cs" />
     <Compile Include="SkippedScenarioFeature.cs" />
     <Compile Include="TeardownFeature.cs" />
     <Compile Include="StepTimeoutFeature.cs" />
